enhance the documentation and formatting remove calculate blob metrics
